114-2/NLP/pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = ""
authors = [
    {name = "JiaLong0209",email = "smt02091395@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "music21 (>=9.9.1,<10.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "google-genai (>=1.66.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
